Move EnumerableExtensions.LongSkip to Libplanet.RocksDBStore
