+++ Libplanet/EnumerableExtensions.cs
@@ -1,3 +1,4 @@
+#if !NETSTANDARD2_1
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
@@ -5,9 +6,8 @@ using System.Linq;
 
 namespace Libplanet
 {
-    public static class EnumerableExtensions
+    internal static class EnumerableExtensions
     {
-#if !NETSTANDARD2_1
         // Copied from .NET Runtime <https://git.io/JL8X6#L115-L154>
         public static IEnumerable<TSource> SkipLast<TSource>(
             this IEnumerable<TSource> source,
@@ -20,25 +20,6 @@ namespace Libplanet
             }
 
             return count <= 0 ? source.Skip(0) : SkipLastIterator(source, count);
-        }
-#endif
-
-        public static IEnumerable<TSource> LongSkip<TSource>(
-            this IEnumerable<TSource> source,
-            long count
-        )
-        {
-            if (count < 0)
-            {
-                count = 0;
-            }
-
-            if (count <= int.MaxValue)
-            {
-                return source.Skip((int)count);
-            }
-
-            return SkipIterator(source, count);
         }
 
         private static IEnumerable<TSource> SkipLastIterator<TSource>(
@@ -70,23 +51,6 @@ namespace Libplanet
                 }
             }
         }
-
-        private static IEnumerable<TSource> SkipIterator<TSource>(
-            IEnumerable<TSource> source,
-            long count
-        )
-        {
-            foreach (var item in source)
-            {
-                if (count > 0)
-                {
-                    --count;
-                }
-                else
-                {
-                    yield return item;
-                }
-            }
-        }
     }
 }
+#endif
